requirements.txt
flask==2.0.1
flask-cors==3.0.10
torch>=2.0.0
torchvision>=0.15.0
transformers==4.26.0
sentence-transformers==2.2.2
scikit-learn==1.0.2
google-generativeai==0.3.1
python-dotenv==0.21.0
numpy==1.24.0
